backend/requirements.txt
fastapi==0.105.0
uvicorn==0.24.0
pydantic==2.5.2
pandas==2.1.3
numpy==1.26.2
scipy==1.11.4
python-dotenv==1.0.0
openai==0.28.0
python-multipart==0.0.6
httpx==0.25.2
matplotlib==3.8.2
starlette==0.27.0 